change tracks duration to integer

+++ sql_scripts/create_tables.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS tracks (
     id VARCHAR(23) NOT NULL,
     name VARCHAR NOT NULL,
     artist_id VARCHAR(23) NOT NULL,
-    duration_ms SMALLINT NOT NULL,
+    duration_ms INTEGER NOT NULL,
     popularity SMALLINT NOT NULL,
     is_in_album BOOLEAN,
     is_explicit BOOLEAN,
